+++ test/noop_implementation.js
@@ -7,41 +7,13 @@ var expect = require('chai').expect;
 // library.  Again, globals are used purely for convenience.
 var opentracing = require('../debug.js');
 
-module.exports = function mockCompatibilityChecks() {
-    describe('Mock Opentracing Implementation', function() {
-        describe('Constants', function() {
-            var constStrings = [
-                'FORMAT_TEXT_MAP',
-                'FORMAT_BINARY',
-                'FORMAT_HTTP_HEADERS',
-                'REFERENCE_CHILD_OF',
-                'REFERENCE_FOLLOWS_FROM',
-            ];
-            _.each(constStrings, function(name) {
-                it(name + ' should be a constant string', function() {
-                    expect(opentracing[name]).to.be.a('string');
-                });
-            });
-        });
+module.exports = function noopImplementationTests(_createTracer) {
+    var createTracer = _createTracer || function() { return new opentracing.Tracer(); };
 
-        describe('Standalone functions', function() {
-            var funcs = [
-                'childOf',
-                'followsFrom',
-                'initGlobalTracer',
-                'globalTracer',
-
-            ];
-            _.each(funcs, function(name) {
-                it(name + ' should be a function', function() {
-                    expect(opentracing[name]).to.be.a('function');
-                });
-            });
-        });
-
+    describe('Noop Tracer Implementation', function() {
         describe('Tracer#inject', function() {
             it('should enforce the required carrier types', function() {
-                var tracer = new opentracing.Tracer();
+                var tracer = createTracer();
                 var spanContext = tracer.startSpan('test_operation').context();
                 expect(function() { tracer.inject(spanContext, opentracing.FORMAT_TEXT_MAP, ''); }).to.throw(Error);
                 expect(function() { tracer.inject(spanContext, opentracing.FORMAT_TEXT_MAP, 5); }).to.throw(Error);
@@ -49,16 +21,24 @@ module.exports = function mockCompatibilityChecks() {
                 expect(function() { tracer.extract(opentracing.FORMAT_BINARY, { buffer : '' }); }).to.throw(Error);
                 expect(function() { tracer.extract(opentracing.FORMAT_BINARY, { buffer : 5 }); }).to.throw(Error);
             });
+
+            it('should handle Spans and SpanContexts', function() {
+                var tracer = createTracer();
+                var span = tracer.startSpan('test_operation');
+                var textCarrier = {};
+                expect(function() { tracer.inject(span, opentracing.FORMAT_TEXT_MAP, textCarrier); }).to.not.throw(Error);
+            });
         });
 
         describe('Span#finish', function() {
             it('should return undefined', function() {
-                var tracer = new opentracing.Tracer();
+                var tracer = createTracer();
                 var span = tracer.startSpan('test_span');
                 expect(span.finish()).to.be.undefined;
             });
+
             it('should only accept numbers as an argument', function() {
-                var tracer = new opentracing.Tracer();
+                var tracer = createTracer();
                 function f(arg) {
                     return function() {
                         var span = tracer.startSpan('test_span');
@@ -70,7 +50,7 @@ module.exports = function mockCompatibilityChecks() {
                 expect(f({})).to.throw(Error);
             });
             it('should throw an Error if given more than 1 argument', function() {
-                var tracer = new opentracing.Tracer();
+                var tracer = createTracer();
                 var span = tracer.startSpan('test_span');
                 expect(function() { span.finish(Date.now(), { extra : 123 }) }).to.be.throw(Error);
             });
@@ -79,7 +59,7 @@ module.exports = function mockCompatibilityChecks() {
         describe('Miscellaneous', function() {
             describe('Memory usage', function() {
                 it('should not report leaks after setting the global tracer', function() {
-                    opentracing.initGlobalTracer(new opentracing.Tracer());
+                    opentracing.initGlobalTracer(createTracer());
                 });
             });
         });
